api.py refactor
